料金計算のJSの記述場所を修正
