+++ app/javascript/packs/application.js
@@ -1,10 +1,18 @@
-window.addEventListener('load', () => {
-  const itemPrice = document.getElementById("item-price");
-  const addTaxPrice = document.getElementById("add-tax-price");
-  const profit = document.getElementById("profit");
+-// This file is automatically compiled by Webpack, along with any other files
+// present in this directory. You're encouraged to place your actual application logic in
+// a relevant structure within app/javascript and only use these pack files to reference
+// that code so it'll be compiled.
 
-  itemPrice.addEventListener("input", () => {
-    addTaxPrice.innerHTML = Math.round(itemPrice.value * 0.1 );
-    profit.innerHTML = Math.round(itemPrice.value - Math.round(itemPrice.value * 0.1 ));
-  });
-});
+require("@rails/ujs").start()
+require("@rails/activestorage").start()
+require("channels")
+require("../item_price")
+
+
+// Uncomment to copy all static images under ../images to the output folder and reference
+// them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
+// or the `imagePath` JavaScript helper below.
+//
+// const images = require.context('../images', true)
+// const imagePath = (name) => images(name, true)
+
